backend/requirements.txt
playwright
google-generativeai
fastapi
uvicorn
websockets
pillow
python-dotenv